refactor: Move parser unit tests from Orgfile to new parse_org test file
